Move event publisher to new service
