+++ service/src/main/java/bio/terra/externalcreds/services/LinkedAccountService.java
@@ -2,8 +2,6 @@ package bio.terra.externalcreds.services;
 
 import bio.terra.common.db.ReadTransaction;
 import bio.terra.common.db.WriteTransaction;
-import bio.terra.externalcreds.ExternalCredsException;
-import bio.terra.externalcreds.config.ExternalCredsConfig;
 import bio.terra.externalcreds.dataAccess.GA4GHPassportDAO;
 import bio.terra.externalcreds.dataAccess.GA4GHVisaDAO;
 import bio.terra.externalcreds.dataAccess.LinkedAccountDAO;
@@ -12,23 +10,11 @@ import bio.terra.externalcreds.models.GA4GHVisa;
 import bio.terra.externalcreds.models.LinkedAccount;
 import bio.terra.externalcreds.models.LinkedAccountWithPassportAndVisas;
 import bio.terra.externalcreds.visaComparators.VisaComparator;
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.google.api.core.ApiFutureCallback;
-import com.google.api.core.ApiFutures;
-import com.google.cloud.pubsub.v1.Publisher;
-import com.google.common.annotations.VisibleForTesting;
-import com.google.common.util.concurrent.MoreExecutors;
-import com.google.protobuf.ByteString;
-import com.google.pubsub.v1.PubsubMessage;
-import java.io.IOException;
 import java.sql.Timestamp;
 import java.util.Collection;
 import java.util.List;
 import java.util.Optional;
-import java.util.concurrent.TimeUnit;
 import java.util.stream.Collectors;
-import javax.annotation.PreDestroy;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
@@ -39,35 +25,20 @@ public class LinkedAccountService {
   private final LinkedAccountDAO linkedAccountDAO;
   private final GA4GHPassportDAO ga4ghPassportDAO;
   private final GA4GHVisaDAO ga4ghVisaDAO;
-  private final ObjectMapper objectMapper;
-  private final Optional<Publisher> authorizationChangeEventPublisher;
   private final Collection<VisaComparator> visaComparators;
+  private final EventPublisher eventPublisher;
 
   public LinkedAccountService(
       LinkedAccountDAO linkedAccountDAO,
       GA4GHPassportDAO ga4ghPassportDAO,
       GA4GHVisaDAO ga4ghVisaDAO,
-      ExternalCredsConfig externalCredsConfig,
-      ObjectMapper objectMapper,
-      Collection<VisaComparator> visaComparators) {
+      Collection<VisaComparator> visaComparators,
+      EventPublisher eventPublisher) {
     this.linkedAccountDAO = linkedAccountDAO;
     this.ga4ghPassportDAO = ga4ghPassportDAO;
     this.ga4ghVisaDAO = ga4ghVisaDAO;
-    this.objectMapper = objectMapper;
-
-    // note that Publisher authenticates to Google using the env var GOOGLE_APPLICATION_CREDENTIALS
-    this.authorizationChangeEventPublisher =
-        externalCredsConfig
-            .getAuthorizationChangeEventTopicName()
-            .map(
-                topicName -> {
-                  try {
-                    return Publisher.newBuilder(topicName).build();
-                  } catch (IOException e) {
-                    throw new ExternalCredsException("exception building event publisher", e);
-                  }
-                });
     this.visaComparators = visaComparators;
+    this.eventPublisher = eventPublisher;
   }
 
   @ReadTransaction
@@ -96,7 +67,7 @@ public class LinkedAccountService {
             linkedAccountWithPassportAndVisas.withLinkedAccount(savedLinkedAccount));
 
     if (authorizationsDiffer(existingVisas, savedLinkedAccountWithPassportAndVisas.getVisas())) {
-      publishAuthorizationChangeEvent(
+      eventPublisher.publishAuthorizationChangeEvent(
           new AuthorizationChangeEvent.Builder()
               .providerId(savedLinkedAccount.getProviderId())
               .userId(savedLinkedAccount.getUserId())
@@ -116,7 +87,7 @@ public class LinkedAccountService {
     var existingVisas = ga4ghVisaDAO.listVisas(userId, providerId);
     var accountExisted = linkedAccountDAO.deleteLinkedAccountIfExists(userId, providerId);
     if (!existingVisas.isEmpty()) {
-      publishAuthorizationChangeEvent(
+      eventPublisher.publishAuthorizationChangeEvent(
           new AuthorizationChangeEvent.Builder().providerId(providerId).userId(userId).build());
     }
     return accountExisted;
@@ -148,48 +119,6 @@ public class LinkedAccountService {
     } else {
       return linkedAccountWithPassportAndVisas;
     }
-  }
-
-  @VisibleForTesting
-  void publishAuthorizationChangeEvent(AuthorizationChangeEvent event) {
-    authorizationChangeEventPublisher.ifPresent(
-        publisher -> {
-          try {
-            var message =
-                PubsubMessage.newBuilder()
-                    .setData(ByteString.copyFromUtf8(objectMapper.writeValueAsString(event)))
-                    .build();
-            var apiFuture = publisher.publish(message);
-            ApiFutures.addCallback(
-                apiFuture,
-                new ApiFutureCallback<>() {
-                  @Override
-                  public void onFailure(Throwable throwable) {
-                    log.error("failure publishing authorization change event", throwable);
-                  }
-
-                  @Override
-                  public void onSuccess(String messageId) {}
-                },
-                MoreExecutors.directExecutor());
-          } catch (JsonProcessingException e) {
-            throw new ExternalCredsException(
-                "json exception writing authorization change event:" + event, e);
-          }
-        });
-  }
-
-  @PreDestroy
-  void shutdownPublisher() {
-    authorizationChangeEventPublisher.ifPresent(
-        publisher -> {
-          try {
-            publisher.shutdown();
-            publisher.awaitTermination(1, TimeUnit.MINUTES);
-          } catch (InterruptedException e) {
-            throw new ExternalCredsException("publisher shutdown interrupted", e);
-          }
-        });
   }
 
   private boolean authorizationsDiffer(
